[refactor] move `bboxes_iou` to utils
